feedback/pyproject.toml
[project]
name = "feedback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "reportlab>=4.0.0",

]